refactor: separate runtime types from DB contracts
Move runtime types (AG-UI events, tool types, capability types) from
everruns-contracts to everruns-core, making core independent of contracts.

Key changes:
- Create ag_ui.rs in core with AG-UI event types
- Create tool_types.rs in core with ToolCall, ToolResult, ToolDefinition
- Create capability_types.rs in core with CapabilityId, CapabilityStatus
- Remove everruns-contracts dependency from everruns-core
- Update contracts to re-export runtime types from core
- Move DbMessageStore from worker to storage
- Update worker to re-export DbMessageStore from storage

Dependency structure:
- core: No dependency on contracts (purely runtime abstractions)
- contracts: Depends on core, re-exports runtime types for compatibility
- storage: Depends on core and contracts, provides DbMessageStore
- worker: Depends on storage for DbMessageStore

This separation ensures:
- Core remains focused on runtime abstractions
- Contracts contains DB/API entity types (with ToSchema for OpenAPI)
- Storage provides trait implementations (MessageStore)
