+++ crates/everruns-core/Cargo.toml
@@ -35,8 +35,8 @@ chrono.workspace = true
 # Tracing
 tracing.workspace = true
 
-# Internal dependencies
-everruns-contracts = { path = "../everruns-contracts" }
+# Note: This crate has NO dependency on storage/contracts - it's purely runtime abstractions
+# AG-UI events and tool types are defined locally in this crate
 
 # OpenAI Protocol provider
 reqwest = { version = "0.12", features = ["json", "stream"] }
